refactor: modularize modal manager
